+++ test/functional/workgroup/tests/test-workgroup-Ordered-MultipleReuse.hpp
@@ -6,16 +6,17 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
 ///
-/// Header file containing tests for RAJA workgroup unordered runs.
+/// Header file containing tests for RAJA workgroup ordered runs.
 ///
 
-#ifndef __TEST_WORKGROUP_UNORDERED__
-#define __TEST_WORKGROUP_UNORDERED__
+#ifndef __TEST_WORKGROUP_ORDERED_MULTIPLEREUSE__
+#define __TEST_WORKGROUP_ORDERED_MULTIPLEREUSE__
 
 #include "RAJA_test-workgroup.hpp"
 #include "RAJA_test-forall-data.hpp"
 
 #include <random>
+#include <vector>
 
 
 template <typename ExecPolicy,
@@ -25,105 +26,7 @@ template <typename ExecPolicy,
           typename Allocator,
           typename WORKING_RES
           >
-void testWorkGroupUnorderedSingle(IndexType begin, IndexType end)
-{
-  using WorkPool_type = RAJA::WorkPool<
-                  RAJA::WorkGroupPolicy<ExecPolicy, OrderPolicy, StoragePolicy>,
-                  IndexType,
-                  RAJA::xargs<>,
-                  Allocator
-                >;
-
-  using WorkGroup_type = RAJA::WorkGroup<
-                  RAJA::WorkGroupPolicy<ExecPolicy, OrderPolicy, StoragePolicy>,
-                  IndexType,
-                  RAJA::xargs<>,
-                  Allocator
-                >;
-
-  using WorkSite_type = RAJA::WorkSite<
-                  RAJA::WorkGroupPolicy<ExecPolicy, OrderPolicy, StoragePolicy>,
-                  IndexType,
-                  RAJA::xargs<>,
-                  Allocator
-                >;
-
-  ASSERT_GE(begin, (IndexType)0);
-  ASSERT_GE(end, begin);
-  IndexType N = end + begin;
-
-  camp::resources::Resource working_res{WORKING_RES::get_default()};
-
-  IndexType* working_array;
-  IndexType* check_array;
-  IndexType* test_array;
-
-  allocateForallTestData<IndexType>(N,
-                                    working_res,
-                                    &working_array,
-                                    &check_array,
-                                    &test_array);
-
-
-  {
-    for (IndexType i = IndexType(0); i < N; i++) {
-      test_array[i] = IndexType(0);
-    }
-
-    working_res.memcpy(working_array, test_array, sizeof(IndexType) * N);
-
-    for (IndexType i = begin; i < end; ++i) {
-      test_array[ i ] = IndexType(i);
-    }
-  }
-
-  WorkPool_type pool(Allocator{});
-
-  IndexType test_val(5);
-
-  {
-    pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin, end },
-        [=] RAJA_HOST_DEVICE (IndexType i) {
-      working_array[i] += i + test_val;
-    });
-  }
-
-  WorkGroup_type group = pool.instantiate();
-
-  WorkSite_type site = group.run();
-
-  working_res.memcpy(check_array, working_array, sizeof(IndexType) * N);
-
-  {
-    working_res.memcpy(check_array, working_array, sizeof(IndexType) * N);
-
-    for (IndexType i = IndexType(0); i < begin; i++) {
-      ASSERT_EQ(test_array[i], check_array[i]);
-    }
-    for (IndexType i = begin;        i < end;   i++) {
-      ASSERT_EQ(test_array[i] + test_val, check_array[i]);
-    }
-    for (IndexType i = end;          i < N;     i++) {
-      ASSERT_EQ(test_array[i], check_array[i]);
-    }
-  }
-
-
-  deallocateForallTestData<IndexType>(working_res,
-                                      working_array,
-                                      check_array,
-                                      test_array);
-}
-
-
-template <typename ExecPolicy,
-          typename OrderPolicy,
-          typename StoragePolicy,
-          typename IndexType,
-          typename Allocator,
-          typename WORKING_RES
-          >
-void testWorkGroupUnorderedMultiple(
+void testWorkGroupOrderedMultiple(
     std::mt19937& rng, IndexType max_begin, IndexType min_end,
     IndexType num1, IndexType num2, IndexType num3,
     IndexType pool_reuse, IndexType group_reuse)
@@ -175,6 +78,7 @@ void testWorkGroupUnorderedMultiple(
     }
   }
 
+
   camp::resources::Resource working_res{WORKING_RES::get_default()};
 
   using type1 = IndexType;
@@ -213,6 +117,8 @@ void testWorkGroupUnorderedMultiple(
 
 
   WorkPool_type pool(Allocator{});
+  WorkGroup_type group = pool.instantiate();
+  WorkSite_type site = group.run();
 
   for (IndexType pr = 0; pr < pool_reuse; pr++) {
 
@@ -226,7 +132,11 @@ void testWorkGroupUnorderedMultiple(
         type1* working_ptr1 = working_array1 + N * j;
         pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin1[j], end1[j] },
             [=] RAJA_HOST_DEVICE (IndexType i) {
-          working_ptr1[i] += type1(i) + test_val1;
+          working_ptr1[i] += type1(i);
+        });
+        pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin1[j], end1[j] },
+            [=] RAJA_HOST_DEVICE (IndexType i) {
+          working_ptr1[i] += test_val1;
         });
       }
 
@@ -234,7 +144,11 @@ void testWorkGroupUnorderedMultiple(
         type2* working_ptr2 = working_array2 + N * j;
         pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin2[j], end2[j] },
             [=] RAJA_HOST_DEVICE (IndexType i) {
-          working_ptr2[i] += type2(i) + test_val2;
+          working_ptr2[i] += type2(i);
+        });
+        pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin2[j], end2[j] },
+            [=] RAJA_HOST_DEVICE (IndexType i) {
+          working_ptr2[i] += test_val2;
         });
       }
 
@@ -242,12 +156,16 @@ void testWorkGroupUnorderedMultiple(
         type3* working_ptr3 = working_array3 + N * j;
         pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin3[j], end3[j] },
             [=] RAJA_HOST_DEVICE (IndexType i) {
-          working_ptr3[i] += type3(i) + test_val3;
+          working_ptr3[i] += type3(i);
+        });
+        pool.enqueue(RAJA::TypedRangeSegment<IndexType>{ begin3[j], end3[j] },
+            [=] RAJA_HOST_DEVICE (IndexType i) {
+          working_ptr3[i] += test_val3;
         });
       }
     }
 
-    WorkGroup_type group = pool.instantiate();
+    group = pool.instantiate();
 
     for (IndexType gr = 0; gr < group_reuse; gr++) {
 
@@ -304,7 +222,7 @@ void testWorkGroupUnorderedMultiple(
         }
       }
 
-      WorkSite_type site = group.run();
+      site = group.run();
 
       // check_test_data(type1(5), type2(7), type3(11));
       {
@@ -359,6 +277,8 @@ void testWorkGroupUnorderedMultiple(
       }
     }
 
+    site.clear();
+    group.clear();
     pool.clear();
   }
 
@@ -381,47 +301,14 @@ void testWorkGroupUnorderedMultiple(
 
 
 template <typename T>
-class WorkGroupBasicUnorderedSingleFunctionalTest : public ::testing::Test
+class WorkGroupBasicOrderedMultipleReuseFunctionalTest : public ::testing::Test
 {
 };
 
-TYPED_TEST_SUITE_P(WorkGroupBasicUnorderedSingleFunctionalTest);
-
-template <typename T>
-class WorkGroupBasicUnorderedMultipleReuseFunctionalTest : public ::testing::Test
-{
-};
-
-TYPED_TEST_SUITE_P(WorkGroupBasicUnorderedMultipleReuseFunctionalTest);
+TYPED_TEST_SUITE_P(WorkGroupBasicOrderedMultipleReuseFunctionalTest);
 
 
-TYPED_TEST_P(WorkGroupBasicUnorderedSingleFunctionalTest, BasicWorkGroupUnorderedSingle)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using OrderPolicy = typename camp::at<TypeParam, camp::num<1>>::type;
-  using StoragePolicy = typename camp::at<TypeParam, camp::num<2>>::type;
-  using IndexType = typename camp::at<TypeParam, camp::num<3>>::type;
-  using Allocator = typename camp::at<TypeParam, camp::num<4>>::type;
-  using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<5>>::type;
-
-  std::mt19937 rng(std::random_device{}());
-  using dist_type = std::uniform_int_distribution<IndexType>;
-
-  IndexType b1 = dist_type(IndexType(0), IndexType(15))(rng);
-  IndexType e1 = dist_type(b1, IndexType(16))(rng);
-
-  IndexType b2 = dist_type(e1, IndexType(127))(rng);
-  IndexType e2 = dist_type(b2, IndexType(128))(rng);
-
-  IndexType b3 = dist_type(e2, IndexType(1023))(rng);
-  IndexType e3 = dist_type(b3, IndexType(1024))(rng);
-
-  testWorkGroupUnorderedSingle< ExecPolicy, OrderPolicy, StoragePolicy, IndexType, Allocator, WORKING_RESOURCE >(b1, e1);
-  testWorkGroupUnorderedSingle< ExecPolicy, OrderPolicy, StoragePolicy, IndexType, Allocator, WORKING_RESOURCE >(b2, e2);
-  testWorkGroupUnorderedSingle< ExecPolicy, OrderPolicy, StoragePolicy, IndexType, Allocator, WORKING_RESOURCE >(b3, e3);
-}
-
-TYPED_TEST_P(WorkGroupBasicUnorderedMultipleReuseFunctionalTest, BasicWorkGroupUnorderedMultipleReuse)
+TYPED_TEST_P(WorkGroupBasicOrderedMultipleReuseFunctionalTest, BasicWorkGroupOrderedMultipleReuse)
 {
   using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
   using OrderPolicy = typename camp::at<TypeParam, camp::num<1>>::type;
@@ -440,8 +327,8 @@ TYPED_TEST_P(WorkGroupBasicUnorderedMultipleReuseFunctionalTest, BasicWorkGroupU
   IndexType pool_reuse  = dist_type(IndexType(0), IndexType(8))(rng);
   IndexType group_reuse = dist_type(IndexType(0), IndexType(8))(rng);
 
-  testWorkGroupUnorderedMultiple< ExecPolicy, OrderPolicy, StoragePolicy, IndexType, Allocator, WORKING_RESOURCE >(
+  testWorkGroupOrderedMultiple< ExecPolicy, OrderPolicy, StoragePolicy, IndexType, Allocator, WORKING_RESOURCE >(
       rng, IndexType(96), IndexType(4000), num1, num2, num3, pool_reuse, group_reuse);
 }
 
-#endif  //__TEST_WORKGROUP_UNORDERED__
+#endif  //__TEST_WORKGROUP_ORDERED_MULTIPLEREUSE__
